+++ Core/include/Acts/Material/MaterialSlab.hpp
@@ -45,12 +45,12 @@ class MaterialSlab {
   /// Construct vacuum without thickness.
   MaterialSlab() = default;
   /// Construct vacuum with thickness.
-  explicit MaterialSlab(float thickness);
+  explicit MaterialSlab(double thickness);
   /// Construct from material description.
   ///
   /// @param material  is the material description
   /// @param thickness is the thickness of the material
-  MaterialSlab(const Material& material, float thickness);
+  MaterialSlab(const Material& material, double thickness);
   ~MaterialSlab() = default;
 
   MaterialSlab(MaterialSlab&&) = default;
@@ -59,7 +59,7 @@ class MaterialSlab {
   MaterialSlab& operator=(const MaterialSlab&) = default;
 
   /// Scale the material thickness by the given factor.
-  void scaleThickness(float scale);
+  void scaleThickness(double scale);
 
   /// Check if the material is valid, i.e. it is finite and not vacuum.
   bool isValid() const { return m_material.isValid() && (0.0f < m_thickness); }
@@ -67,17 +67,17 @@ class MaterialSlab {
   /// Access the (average) material parameters.
   constexpr const Material& material() const { return m_material; }
   /// Return the thickness.
-  constexpr float thickness() const { return m_thickness; }
+  constexpr double thickness() const { return m_thickness; }
   /// Return the radiation length fraction.
-  constexpr float thicknessInX0() const { return m_thicknessInX0; }
+  constexpr double thicknessInX0() const { return m_thicknessInX0; }
   /// Return the nuclear interaction length fraction.
-  constexpr float thicknessInL0() const { return m_thicknessInL0; }
+  constexpr double thicknessInL0() const { return m_thicknessInL0; }
 
  private:
   Material m_material;
-  float m_thickness = 0.0f;
-  float m_thicknessInX0 = 0.0f;
-  float m_thicknessInL0 = 0.0f;
+  double m_thickness = 0.0f;
+  double m_thicknessInX0 = 0.0f;
+  double m_thicknessInL0 = 0.0f;
 
   friend constexpr bool operator==(const MaterialSlab& lhs,
                                    const MaterialSlab& rhs) {
